requirements.txt
langchain
openai
faiss-cpu
PyMuPDF
streamlit
python-dotenv
tqdm
sentence-transformers
langchain_community
tiktoken
sentence-transformers